Changes adapted to new structure
