+++ src/main/asciidocs/dataobjects.adoc
@@ -839,6 +839,12 @@ Set the trust options in jks format, aka Java truststore
 |[[tryUseCompression]]`@tryUseCompression`|`Boolean`|+++
 Set whether compression is enabled
 +++
+|[[tryUsePerFrameWebsocketCompression]]`@tryUsePerFrameWebsocketCompression`|`Boolean`|+++
+Set option to offer Deflate Frame websocket compression
++++
+|[[tryUsePerMessageWebsocketCompression]]`@tryUsePerMessageWebsocketCompression`|`Boolean`|+++
+Set option to offer Permessage Deflate websocket compression
++++
 |[[useAlpn]]`@useAlpn`|`Boolean`|+++
 Set the ALPN usage.
 +++
@@ -847,6 +853,15 @@ Set whether Netty pooled buffers are enabled
 +++
 |[[verifyHost]]`@verifyHost`|`Boolean`|+++
 Set whether hostname verification is enabled
++++
+|[[websocketCompressionAllowClientNoContext]]`@websocketCompressionAllowClientNoContext`|`Boolean`|+++
+Set the websocket compression allow client no context option
++++
+|[[websocketCompressionLevel]]`@websocketCompressionLevel`|`Number (int)`|+++
+Set websocket compression level
++++
+|[[websocketCompressionRequestServerNoContext]]`@websocketCompressionRequestServerNoContext`|`Boolean`|+++
+Set the websocket compression server no context option
 +++
 |===
 
@@ -978,6 +993,12 @@ Set the key/cert store options in pem format.
 |[[pemTrustOptions]]`@pemTrustOptions`|`link:dataobjects.html#PemTrustOptions[PemTrustOptions]`|+++
 Set the trust options in pem format
 +++
+|[[perFrameWebsocketCompressionSupported]]`@perFrameWebsocketCompressionSupported`|`Boolean`|+++
+Enable or disable support for WebSocket Defalte Frame compression
++++
+|[[perMessageWebsocketCompressionSupported]]`@perMessageWebsocketCompressionSupported`|`Boolean`|+++
+Enable or disable support for WebSocket Permessage Deflate compression
++++
 |[[pfxKeyCertOptions]]`@pfxKeyCertOptions`|`link:dataobjects.html#PfxOptions[PfxOptions]`|+++
 Set the key/cert options in pfx format.
 +++
@@ -1036,6 +1057,15 @@ Set the ALPN usage.
 +++
 |[[usePooledBuffers]]`@usePooledBuffers`|`Boolean`|+++
 Set whether Netty pooled buffers are enabled
++++
+|[[websocketAllowServerNoContext]]`@websocketAllowServerNoContext`|`Boolean`|+++
+Set the WebSocket Allow Server No Context option
++++
+|[[websocketCompressionLevel]]`@websocketCompressionLevel`|`Number (int)`|+++
+Set the WebSocket compression level
++++
+|[[websocketPreferredClientNoContext]]`@websocketPreferredClientNoContext`|`Boolean`|+++
+Set the WebSocket Preferred Client No Context setting
 +++
 |[[websocketSubProtocols]]`@websocketSubProtocols`|`String`|+++
 Set the websocket subprotocols supported by the server.
